+++ docs/src/changelog.md
@@ -26,6 +26,7 @@ If your currently installed *Quality-time* version is not v5.14.0, please first 
 ### Changed
 
 - Always show metric trend graph tabs. Display a loader in the tab while loading measurements. If the trend graph cannot be displayed, explain in the tab why not. Closes [#8825](https://github.com/ICTU/quality-time/issues/8825).
+- Always show measurement entity tabs. Display a loader in the tab while loading measurements. If the measurement entities cannot be displayed, explain in the tab why not. Closes [#8826](https://github.com/ICTU/quality-time/issues/8826).
 
 ## v5.14.0 - 2024-07-05